chore: tests and bun migration complete phase
