+++ .github/workflows/tests.yml
@@ -1,5 +1,5 @@
-# This workflow will do a clean install of node dependencies, build the source code and run tests across different versions of node
-# For more information see: https://help.github.com/actions/language-and-framework-guides/using-nodejs-with-github-actions
+# This workflow will do a clean install of dependencies, build the source code and run tests
+# For more information see: https://github.com/oven-sh/setup-bun
 
 name: build and test
 
@@ -9,15 +9,11 @@ jobs:
   build-and-test:
     runs-on: ubuntu-latest
 
-    strategy:
-      matrix:
-        node-version: [lts/*]
-
     steps:
-      - uses: actions/checkout@v2
-      - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v4
+      - uses: actions/checkout@v4
+      - name: Use Bun
+        uses: oven-sh/setup-bun@v2
         with:
-          node-version: ${{ matrix.node-version }}
-      - run: npm clean-install
-      - run: npm test
+          bun-version-file: package.json
+      - run: bun install --frozen-lockfile
+      - run: bun run test
