+++ examples/example-projects/plain-javascript-modules/README.md
@@ -2,6 +2,6 @@ This example is written in JavaScript without any build, using JavaScript
 modules to organize code into separate files and importing from one file to the
 other using native `import` syntax.
 
-To set up, run `npm install`.
+To set up, run `bun install`.
 
-To run the project, run `npm start`.
+To run the project, run `bun start`.
