+++ package.json
@@ -37,14 +37,14 @@
 		"build": "bun run scripts/write-version.ts && bun run scripts/build.ts",
 		"examples": "bunx serve .",
 		"test": "bun run build && tsc && bun run test-lint && bun run test-unit",
-		"test-unit": "nodeunit test/unit/nodeunitheadless.cjs",
+		"test-unit": "bun test",
 		"test-lint": "biome check .",
 		"lint": "prettier . --write",
 		"prepare22": "bun run build",
 		"version": "bun test && git add .",
-		"release:patch": "npm version patch --message 'v%s' && npm publish && npm run _release:push-branch",
-		"release:minor": "npm version minor --message 'v%s' && npm publish && npm run _release:push-branch",
-		"release:major": "npm version major --message 'v%s' && npm publish && npm run _release:push-branch",
+		"release:patch": "npm version patch --message 'v%s' && npm publish && bun run _release:push-branch",
+		"release:minor": "npm version minor --message 'v%s' && npm publish && bun run _release:push-branch",
+		"release:major": "npm version major --message 'v%s' && npm publish && bun run _release:push-branch",
 		"_release:push-branch": "git push --follow-tags --set-upstream origin `git rev-parse --abbrev-ref HEAD`"
 	},
 	"author": "tween.js contributors (https://github.com/tweenjs/tween.js/graphs/contributors)",
