--- test/unit/nodeunit.html
@@ -1,14 +1,0 @@
-<html>
-	<head>
-		<title>nodeunit based tests</title>
-		<script src="nodeunit.js"></script>
-		<script src="../../dist/tween.umd.js"></script>
-		<!-- Run the build first for this file to exist. -->
-		<script src="../../.tmp/tests.umd.js"></script>
-	</head>
-	<body>
-		<script>
-			nodeunit.run({tween: TWEEN.tests})
-		</script>
-	</body>
-</html>
