--- test/unit/nodeunitheadless.cjs
@@ -1,4 +1,0 @@
-const {tests} = require('../../.tmp/tests.cjs')
-module.exports = {tween: tests}
-
-// CONTINUE: got html tests running, had to convert to CJS files, ammend the infrastructure update commit with these changes
