+++ tsconfig.build.json
@@ -1,0 +1,12 @@
+{
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"noEmit": false,
+		"declaration": true,
+		"emitDeclarationOnly": true,
+		"rootDir": "./src",
+		"outDir": "./dist/types"
+	},
+	"include": ["src"],
+	"exclude": ["src/global.ts"]
+}
