+++ tsconfig.json
@@ -8,6 +8,6 @@
 		"types": ["bun"],
 		"noEmit": true
 	},
-	"include": ["src", "scripts"],
+	"include": ["src", "scripts", "test"],
 	"exclude": ["node_modules", "examples"]
 }
